jatek logging

+++ Jatek/sd.Jatek.Application/CommandHandlers/RemovePlayerCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using MediatR;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using sd.Jatek.Application.Commands;
 using sd.Jatek.Infrastructure;
 
@@ -7,17 +9,30 @@ namespace sd.Jatek.Application.CommandHandlers
     public class RemovePlayerCommandHandler : IRequestHandler<RemovePlayerCommand>
     {
         private readonly GameContext _context;
-        public RemovePlayerCommandHandler(GameContext context)
+        private readonly ILogger<RemovePlayerCommandHandler> _logger;
+
+        public RemovePlayerCommandHandler(GameContext context, ILogger<RemovePlayerCommandHandler> logger)
         {
             _context = context;
+            _logger = logger;
         }
+
         public async Task<Unit> Handle(RemovePlayerCommand request, CancellationToken cancellationToken)
         {
-            var player = _context.Players.FirstOrDefault(p => p.PlayerId == request.PlayerId);
+            var player = await _context.Players
+                .FirstOrDefaultAsync(p => p.PlayerId == request.PlayerId, cancellationToken);
 
-            _context.Players.Remove(player);
+            if (player != null)
+                _context.Players.Remove(player);
+            else
+                _logger.LogError("Player with id: {playerId} doesn't exist",
+                request.PlayerId);
 
             await _context.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Player {playerName} was removed from room: {roomId}",
+                player?.PlayerName,
+                player?.RoomId);
 
             return Unit.Value;
         }
